Added optimizations when rawContent is saved
